pyproject.toml
[project]
name = "hello-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ZaidFareed", email = "zaydfareed094@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.302",
    "openai-agents>=0.0.7",
]

[project.scripts]
hello-chainlit = "hello_chainlit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
